Enable network diagnostics for Gen3 devices
